Feature/view list weather

+++ weather-app/Network/Weather/ResponseModel/MainResponseModel.swift
@@ -1,10 +1,10 @@
 import Foundation
 
 struct MainResponseModel: Codable {
-    let temp: Float
-    let feelsLike: Float
-    let tempMin: Float
-    let tempMax: Float
-    let pressure: Float
-    let humidity: Float
+    let temp: Double
+    let feelsLike: Double
+    let tempMin: Double
+    let tempMax: Double
+    let pressure: Double
+    let humidity: Double
 }
